+++ examples/src/linearlite/src/context/MenuContext.ts
@@ -1,8 +1,0 @@
-import React from 'react'
-
-interface MenuContextInterface {
-  showMenu: boolean
-  setShowMenu: (show: boolean) => void
-}
-
-export const MenuContext = React.createContext(null as MenuContextInterface | null)